web/backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.0
pyyaml>=6.0
